requirements.txt
dj-database-url==0.4.2
Django==1.9.2
django-ajax-selects==1.5.2
django-countries==3.4.1
django-dotenv==1.4.1
django-getenv==1.3.1
git+https://github.com/Work4Labs/django-test-pep8
djangorestframework==3.6.2
pep8==1.7.0
pyldap==2.4.28
